+++ changelog.md
@@ -14,6 +14,7 @@
 * LLT-4490: Ensure that meshnet entities are started only when meshnet is started
 * LLT-4335: Make blocking upnp gw search async
 * LLT-4620: Implement the post quantum key rotation mechanism
+* LLT-4616: Make nurse qos ping asynchronous
 
 <br>